[Feat] Log 생성 API 구현

+++ src/main/java/org/moonshot/server/domain/keyresult/model/KeyResult.java
@@ -26,7 +26,7 @@ public class KeyResult {
     private Period period;
 
     @Column(nullable = false)
-    private Integer target;
+    private Long target;
 
     @Column(nullable = false)
     private Short idx;
@@ -64,7 +64,7 @@ public class KeyResult {
         this.idx = idx;
     }
 
-    public void modifyTarget(Integer target) {
+    public void modifyTarget(Long target) {
         this.target = target;
     }
